Added a way to write tests for most of the functions. mkdir, since it has side effects, isn't very well suited but we have a kind of test for it anyhow.
I made a big change to the $(equals) function - if both arguments are
empty it returns a '1' to indicate that they were equal. This may not be
ideal but it's all I could think of as a way of ensuring that an $(if)
statement will work properly with $(equals).

The tests for the comparison operators don't all work yet because those
functions fail if their arguments have excess whitespace. I think this
is too harsh and I'll fix it in a following commit.